CBB Analytics/requirements.txt
playwright>=1.49.0
pandas>=2.1.3
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
lxml>=4.9.0
